+++ app/src/main/java/com/softdata/dyhxx/helper/db/dataRepository/ICarRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface ICarRepository {
     suspend fun insertCar(carEntity: CarEntity): Long
 
-    suspend fun deleteCar(id: Long): Int
+    suspend fun deleteCar(id: Int): Int
 
     fun getCar(id: Long): Flow<CarEntity>
